+++ multiversx_sdk_cli/sign_verify.py
